+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * Make preserve author of issue on copy
 * Add relation between Issue and WeeklySchedule 
 * Add issue associations to copy in settings
-* If issue have due_date then copied issue have relative due_date 
+* If issue have due_date then copied issue have relative due_date
+* Add watchers to copied issue
 
 # 0.0.1
 * Add ability to make issues copies